requirements.txt
apiview==1.3.25
arabic-reshaper==2.1.1
asgiref==3.3.1
boto3==1.17.49
botocore==1.20.49
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
databases==0.4.1
decorator==5.0.6
decorators==2.0.7
defusedxml==0.6.0
diff-match-patch==20200713
dj-config-url==0.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.6
django-allauth==0.44.0
django-crispy-forms==1.11.2
django-discover-runner==1.0
django-heroku==0.3.1
django-import-export==2.5.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-role-permissions==3.1.1
django-staff-required==2020.7.1
django-toolbelt==0.0.1
djangorestframework==3.12.2
docs==0.1.0
drf-yasg==1.20.0
dynaconf==3.1.4
environ==1.0
et-xmlfile==1.0.1
future==0.18.2
GitPrompt==0.1.1
gunicorn==20.0.4
html5lib==1.1
hvac==0.10.9
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
jmespath==0.10.0
knox==0.1.11
lockfile==0.12.2
loguru==0.5.3
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==1.0.2
mysqlclient==2.0.3
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.6
packaging==20.9
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
reportlab==3.5.64
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
s3transfer==0.3.6
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
tst==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
utils==1.0.1
validators==0.18.2
webencodings==0.5.1
whitenoise==5.2.0
win32-setctime==1.0.3
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
